# backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "YCulture Quiz Game API",
    "version": "1.0.0",
    "description": "Real-time multiplayer culture quiz game API with WebSocket support",
    "contact": {
      "name": "YCulture Team",
      "email": "support@yculture.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.yculture.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "displayName": { "type": "string" },
          "isAnonymous": { "type": "boolean" },
          "avatarUrl": { "type": "string", "nullable": true },
          "totalGamesPlayed": { "type": "integer" },
          "totalWins": { "type": "integer" },
          "totalScore": { "type": "integer" },
          "highestScore": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "lastSeenAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "Game": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "roomCode": { "type": "string" },
          "hostId": { "type": "integer" },
          "status": { 
            "type": "string",
            "enum": ["waiting", "in_progress", "completed", "cancelled"]
          },
          "mode": {
            "type": "string",
            "enum": ["multiplayer", "solo_practice"]
          },
          "maxPlayers": { "type": "integer" },
          "currentPlayers": { "type": "integer" },
          "questionsPerGame": { "type": "integer" },
          "timePerQuestion": { "type": "integer" },
          "currentQuestionIndex": { "type": "integer" },
          "startedAt": { "type": "string", "format": "date-time", "nullable": true },
          "completedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "LeaderboardEntry": {
        "type": "object",
        "properties": {
          "userId": { "type": "integer" },
          "displayName": { "type": "string" },
          "avatarUrl": { "type": "string", "nullable": true },
          "totalScore": { "type": "integer" },
          "totalWins": { "type": "integer" },
          "totalGamesPlayed": { "type": "integer" },
          "rank": { "type": "integer" }
        }
      },
      "UserStats": {
        "type": "object",
        "properties": {
          "totalGamesPlayed": { "type": "integer", "description": "Total games played" },
          "totalWins": { "type": "integer", "description": "Games won" },
          "totalScore": { "type": "integer", "description": "Total score" },
          "highestScore": { "type": "integer", "description": "Best score" },
          "averageScore": { "type": "number", "format": "float", "description": "Average score per game" },
          "accuracy": { "type": "number", "format": "float", "description": "Answer accuracy %" }
        }
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "questionText": { "type": "string" },
          "questionType": { 
            "type": "string",
            "enum": ["multiple_choice", "true_false"]
          },
          "difficulty": { 
            "type": "string",
            "enum": ["easy", "medium", "hard"]
          },
          "points": { "type": "integer" },
          "timeLimit": { "type": "integer", "description": "Seconds" },
          "imageUrl": { "type": "string", "nullable": true },
          "answers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "answerText": { "type": "string" },
                "displayOrder": { "type": "integer" }
              }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "type": "object" }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register new user",
        "description": "Create a new registered user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "username", "password", "displayName"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "username": { "type": "string", "minLength": 3, "maxLength": 50 },
                  "password": { "type": "string", "minLength": 6 },
                  "displayName": { "type": "string", "minLength": 2, "maxLength": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "$ref": "#/components/schemas/User" },
                        "token": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or user already exists",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login",
        "description": "Authenticate and get access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "$ref": "#/components/schemas/User" },
                        "token": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/auth/anonymous": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Create anonymous user",
        "description": "Create a temporary anonymous user (no registration required)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["displayName"],
                "properties": {
                  "displayName": { "type": "string", "minLength": 2, "maxLength": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Anonymous user created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "$ref": "#/components/schemas/User" },
                        "token": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/games/create": {
      "post": {
        "tags": ["Games"],
        "summary": "Create new game",
        "description": "Create a new game room",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["mode"],
                "properties": {
                  "mode": {
                    "type": "string",
                    "enum": ["multiplayer", "solo_practice"]
                  },
                  "maxPlayers": { "type": "integer", "minimum": 2, "maximum": 20 },
                  "questionsPerGame": { "type": "integer", "minimum": 5, "maximum": 50 },
                  "timePerQuestion": { "type": "integer", "minimum": 10, "maximum": 60 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Game created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Game" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/games/{roomCode}": {
      "get": {
        "tags": ["Games"],
        "summary": "Get game by room code",
        "description": "Retrieve game information using room code",
        "parameters": [
          {
            "name": "roomCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "6-character room code"
          }
        ],
        "responses": {
          "200": {
            "description": "Game found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Game" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/games/{gameId}/start": {
      "post": {
        "tags": ["Games"],
        "summary": "Start game",
        "description": "Start a game (usually called by host)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Game ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Game started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current user profile",
        "description": "Get authenticated user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "description": "Get any user's public profile",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/leaderboard": {
      "get": {
        "tags": ["Leaderboard"],
        "summary": "Get global leaderboard",
        "description": "Get top players (excludes anonymous users)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 100 },
            "description": "Number of entries to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/LeaderboardEntry" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/leaderboard/rank/{userId}": {
      "get": {
        "tags": ["Leaderboard"],
        "summary": "Get user rank",
        "description": "Get a specific user's rank on the leaderboard",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User rank",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "rank": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Authentication",
      "description": "User authentication, registration, and anonymous access"
    },
    {
      "name": "Games",
      "description": "Game room management, creation, and gameplay"
    },
    {
      "name": "Users",
      "description": "User profiles and statistics"
    },
    {
      "name": "Leaderboard",
      "description": "Global and game-specific rankings"
    },
    {
      "name": "Health",
      "description": "Server health monitoring"
    }
  ]
}